Detect discrepancy between experiment design and expression matrix.
The extract command may fail on some datasets if the experiment design
table contain cell identifiers that do not correspond to a column in the
expression matrix. I was not expecting this to ever happen (and I don't
think it should be happening: where do those cells come from if there
are no reads associated with them?).

Allow the extract command to proceed in this situation by simply
dropping the missing cells, but print a warning that there may be
something wrong with the dataset and let the curator decide what to do
(too bad the curator happens to be me...).