Allow corrections to have a "target".
SCEA and FlyBase may need different corrections (in particular, SCEA is
not interested in the distinction between "embryonic/larval" and "adult"
ontology terms, but FlyBase is).

So we allow a correction dictionary to have a "Target" attribute
indicating whether the corrections are intended for SCEA ("scea") or for
internal consumption by FlyBase ("internal"). SCEA-labelled corrections
are not used for FlyBase-targeted processing and internal-labelled
corrections are not used when generating the files intended for SCEA
curators.