Make sure freshly downloaded datasets are added to the cache.
Make the downloading function return True if the dataset was downloaded,
so that the calling code will append the dataset ID of the downloaded
dataset to the datasets dictionary.

This is a temporary fix only, as for now the function does not really
check whether the dataset was fully and correctly downloaded (it only
checks whether the requested dataset is available on the SCEA).

Closes #1