Allow calling 'select' without filters.
Rework the explorer's select command. Move the main logic to the context
object and keep track and which filters are applied.

Calling the command without any positional argument is now supported and
has the effect of showing the currently applied filter.

Calling the command with only the -c option is equivalent to calling the
clear command (resetting the filter), except that the total number of
cells is shown.

closes #5