Add 'simplified_cell_types' dictionary in dataset JSON file.
Allow the curators to override automatically generated cluster symbols
and titles by adding a 'simplified_cell_types' dictionary to the JSON
object describing the dataset. Each key in that dictionary is a FBbt
label for a cell type and the value is the desired simplified label to
use for the cluster symbol.

close #6