Compute the summarised expression table in streaming mode.
Since some datasets have an expression table just too big to fit in
memory, we compute the summarised version by manually parsing the
MatrixMarket file and doing the needed computations on the fly.

This keeps the amount of required memory minimal, and is even faster
than doing all the computations with the table loaded in memory.

With that method, the E-GEOD-141273 dataset can be processed in 40s and
just 136MB of memory, instead of 65s and 11GB of memory.

The entire FCA 10x dataset now only needs ~750MB, instead of ~64GB.

close #7