Allow multi-valued selectors.
Make it possible to select cells using several values in a single
column, by specifying a list of values instead of a single value for a
selector.

That is, it is now possible to have the following in the JSON file
describing a dataset:

```
"selectors": [ "value1", [ "value2", "value3" ] ]
```

This will select cells that have `value1` in the first selector column,
and either `value2` or `value3` in the second selector column.

close #9