Read tracking configuration in the Tracker object.
Instead of getting the configuration value for the DatasetTracker object
in the main FzoContext, pass the entire configuration to the constructor
for DatasetTracker and let it extract the values it needs.

This is to reinforce locality: the FzoContext object does not need to
know anything about the configuration values needed by the tracker.