Add the SourceDataset object.
This is the first step to a big overhaul of the entire curation module.
The aim is to have a proper object model reflecting the FlyBase dataset
model, and also to have some indirection between the expected format of
the JSON file describing a dataset and the code itself (so that changing
the format can later be done without impacting the code at dozens of
places). Ultimately, this will allow a) assembling datasets made of an
arbitrary number of sub-projects and b) assembling datasets from more
than one source dataset.