Do not output explicitly empty fields.
When producing the proforma, make a distinction between 'unset fields'
and 'empty fields'. If the field is unset, we output the corresponding
field header in the proforma so that the curators can fill in the field
a posteriori if they want/need. If the field is explicitly set to an
empty value, we assume the curators already know the field is supposed
to be empty, so there's no need to produce a line in the proforma for
it.