Move the --with-reads option to the curate command.
Move the --with-reads/--without-reads option from the `curate extract`
subcommand to the `curate` command, so that it applies to all
subcommands.

This is mostly intended so that the `proforma` subcommand can be run
without extracting the number of reads, which is helpful when debugging
as extracting said numbers can take some time with large datasets.