Introduce the ProjectContainer class.
We add a new ProjectContainer class to hold the common functionality
between projects and subprojects. Subprojects are directly represented
by an instance of ProjectContainer, while top-level projects are
represented by an instance of Project, which is derived from
ProjectContainer.

That is, we now have

+- DatasetBase
|
`-+ ProjectContainer
  |
  `- Project

instead of

+- DatasetBase
|
+- Project
|
+- Subproject

This eliminates code duplication between the Project and Subproject
classes.