Correctly handle empty lists of CV terms.
If a CV terms list is not explicitly specified in the specification
file, treat it as absent, and write an empty field in the proforma. If
the list is present but empty, do not output the field at all.

This is the same logic as for string fields, simply transposed to lists.