Allow a Result to cover several Assays.
Do not assume a 1-to-1 relationship between an Assay and a Result.
Instead, allow a Result to be an analysis of several Assays, and a
single Assay to be covered by several Results.

Because there's no longer a direct relation between an Assay and its
Results, Result objects are now associated with Project objects.

It is still assumed that every Assay will have *at least* one Result
covering it (and only it). But this lays the groundwork for allowing to
define more Results (covering more than one Assay), on top of the
per-Assay Result.