Allow to define a single project-wide result.
If the specification for a project (or subproject) contains an entry
named 'result', then we construct an extra result covering all samples
in the (sub)project (in addition to the sample-specific results).

The 'result' entry may be a dictionary, in which 'symbol' and 'title'
keys may be used to override the automatically generated symbol and
title for the result.