Allow overriding stage and sex for multi-assays results.
For results that span several assays, allow to specify a custom
developmental stage instead of using the stage associated to the first
sample. All clusters in those results will then be associated with the
specified stage.

For sex, clusters are associated with the sex of the corresponding
samples if all samples have the same sex. If the result covers samples
of different sex, then clusters are not associated with any sex at all.