Remove the fixscea command.
Corrections intended for the SCEA are now prepared and sent to EBI
curators before the dataset even enters the SCEA pipeline, so they are
no longer managed by fuzzy-onions.

This removes the need for distinguishing between correction sets
intended for FlyBase ("internal"), correction sets intended for the
SCEA, and correction sets intended for both; from now all correction
sets defined in the dataset description file are only intended for
FlyBase.