Add `--since` option to the `store findnew` command.
Allow to restrict the output of the `store findnew` command to datasets
that have been loaded after a specified date.

This will make it possible in the future to delete “old” datasets from
the local store (datasets that have already been processed and that are
no longer needed), without having those datasets being listed as “new”
just because they are absent from the local store.