Allow to exclude all clusters from a sample-specific analysis.
For some datasets (e.g. E-GEOD-136172), only the consolidated analysis
performed on the combined samples is really relevant. For these cases,
we allow to exclude all clusters from the per-sample analyses, by adding
a `include_clusters = false` to the sample description (the default
value is True, indicating that clusters for this sample are included in
the analysis).