Update README and provide sample configuration file.
The README file was slowly become obsolete, so update it to reflect the
current behaviour of the program. In particular, remove completely the
section on the JSON description file, which was hopelessly obsolete. A
complete documentation of the description files and how to use them
still needs to be written, but for now, the sample files in the docs
directory will have to suffice.

Also add a sample configuration file in the docs directory. Again, a
proper documentation for that file is missing.