Do not write the grep-svm.sh script on the remote host.
When searching the full-texts archive, do not write the searching script
to a temporary file on the machine hosting the archive. Send the script
directly to the standard input of the remote bash process instead.

This removes the need for an initialisation step (copying the script
over to the remote machine) and a cleaning step (removing the script at
the end), and therefore removes the need for a distinct TextMiner class
entirely.