Show FBrf in tracker listing.
When listing datasets in long format, have a column showing the FBrf
corresponding to the dataset.

Showing a text citation ("Author et al., YYYY") would be better but
would require fetching author data from the database, as the tracking
file does not store this information. Not worth it for now.