Check for non-gene objects in the expression tables.
Sanity check: When reading the expression tables (the raw table to
extract the read counts, or the normalised table to prepare the
summarised expression table), exclude any row that is not about an
actual gene (with a FBgn identifier).

I've come across at least one dataset that contained a few FBti and
RR_transposable_elements objects. Not sure what this means about the
quality of the analysis.