Fix handling of multi-level project hierarchies.
The logic to obtain the top-level project from an arbitrary deep
subproject is flawed and only works when there is one level of
subprojects. With more than one level, the logic enters an infinite
loop.

We fix that logic by properly walking up the project tree, as should
have been done from the beginning.

The bug was never caugt before because until E-GEOD-218661, we never had
to deal with datasets that require more than one level of subprojects.