Fill LC11m field at the result level.
We will now use the result-level LC11m field to record the analysis
method used to produce the cell-by-gene expression matrix, so that we
can tell which datasets have been analysed in the same way and can
therefore be compared.

For now, since the entire Fuzzy-Onions is still only designed to work
with SCEA datasets, we hardcode the value corresponding to the SCEA
analysis pipeline (FBcv:0009058).