Do grepping internally.
When trying to determine whether a given reference is about scRNAseq,
search for the scRNAseq pattern internally. That is, we perform
everything in the current process: we obtain the path to the fulltext
PDF, extract the text from it, and search for the pattern, all within
the Python code. We do _not_ spawn a process to run a script on the SVM
machine.

This will make it easier to later plug in a LLM-based code in which we
"ask" a LLM some questions about the fulltext of a paper.