Refactor the findnew/checknew commands.
Refactor both the findnew and checknew commands around a new
ResultsTable class that encapsulates all the manipulations of the Pandas
dataframe that contains the list of references along with the associated
data (including the results of the relevance check).

Also, change the way the checknew command works: instead of taking the
list of references to check as arguments on the command line, it now
takes as input the same TSV file used and produced by the findnew
command.