Check for unknown clusters when generating summarised expression table.
Add a --proforma option to the `curate sumexpr` command. The option
expects the filename of the proforma that has been generated for the
same dataset than the one for which we generate the summarised
expression table.

When that option is used, the command will check, once the table has
been produced, that its `sample` column only references clusters that
are described in the proforma, and will print a warning for every
cluster that is _not_ described.