Handle special plural names.
When generating the symbol name and title for a cluster, we need the
plural form of the cell type. For most cell types, this is simply done
by appending a final 's'. But we need to make a special case for the
'yolk nucleus' type, whose plural should be 'yolk nuclei' and certainly
not 'yolk nucleuss'.