Export CV term IDs.
When exporting metadata to JSON, make sure that all references to CV
terms are exported as proper term IDs (e.g. "FBbt:12345678").

For FBcv terms, we already have the ID in the dataset description file:
all references to FBcv terms are in the form "<label> ; <id>"), so we
just need to extract the ID part.

For FBbt/FBdv terms, the description file contains only their labels, so
we need look up the IDs in the corresponding ontologies.