Use qualified accession numbers.
Since not all datasets come from the SCEA, we prefix the accession
numbers pointing to the upstream dataset with a prefix name identifying
the upstream data store.

For the SCEA datasets, the "qualified accession number" is simply the
SCEA accession number prefixed with "SCEA:" (that we add ourselves).

For other datasets, it is up to the curator to provide a qualified
accession number in the dataset source description file.