Make sure to produce CURIEs whenever needed.
CV terms are already obtained in CURIE form (e.g. FBbt:12345678), but
identifiers for other types of FlyBase objects (e.g. datasets or genes)
are not (e.g. FBlc1234567), so we prefix them with a "FlyBase:" prefix.

Not sure if we should instead use the FBxx string as the prefix? E.g.
FBlc1234567 -> FBlc:1234567. To check with other FlyBase/Alliance
developers.