Do not assume all datasets come from the EBI SCEA.
While most datasets (in fact, all but _one_, as of August 2025) come
from the EBI Single Cell Expression Atlas, we should not assume this is
always the case. We already have one non-SCEA dataset, and we may have
more such datasets in the future.

We detect whether a dataset is a SCEA dataset or not (by looking at the
accession number in the sources), and only if it is, then:

* we set the LC11m field for the analysis objects to the FBcv term
  representing the SCEA standard analysis pipeline;
* we add the link to the SCEA in the generated proforma.