Implement mappings from FBcv to Alliance terms.
Include the SSSOM mapping set from FBcv as an embedded resource. The set
is slightly modified from the version published by FBcv to account for
the fact that the current version of SSSOM-Py cannot accept literal
mappings.

Use that set to translate:

* project-level FBcv terms to "Data Set Category Tags";
* sample type FBcv terms to OBI terms.

No need to use the mapping set to translate the sample-level assay type
to MMO terms, since the only MMO term we can ever use if MMO:0000862
anyway.